ojoi/missing-record-redirect-query
